WIP: make filters test module
